requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17
pydantic[email]==2.10.3
pydantic-settings==2.6.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
python-dotenv==1.0.1